+++ src/scopes.rs
@@ -1,10 +1,17 @@
 //! Module for all possible scopes in twitch.
-
 use serde::{Deserialize, Serialize};
 use std::borrow::Cow;
 
 macro_rules! scope_impls {
-    ($($i:ident,scope: $rename:literal, doc: $doc:literal);* $(;)? ) => {
+    (@omit #[deprecated($depr:tt)] $i:ident) => {
+        #[cfg(_internal_never)]
+        Self::$i
+    };
+    (@omit $i:ident) => {
+        Self::$i
+    };
+
+    ($($(#[cfg(($cfg:meta))])* $(#[deprecated($depr:meta)])? $i:ident,scope: $rename:literal, doc: $doc:literal);* $(;)? ) => {
         #[doc = "Scopes for twitch."]
         #[doc = ""]
         #[doc = "<https://dev.twitch.tv/docs/authentication/#scopes>"]
@@ -14,6 +21,8 @@ macro_rules! scope_impls {
         #[serde(into = "String")]
         pub enum Scope {
             $(
+                $(#[cfg($cfg)])*
+                $(#[deprecated($depr)])*
                 #[doc = $doc]
                 #[doc = "\n\n"]
                 #[doc = "`"]
@@ -28,9 +37,12 @@ macro_rules! scope_impls {
 
         impl std::fmt::Display for Scope {
             fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                #![allow(deprecated)]
+
                 f.write_str(match self {
                     Scope::Other(s) => &s,
                     $(
+                        $(#[cfg($cfg)])*
                         Scope::$i => $rename,
                     )*
                 })
@@ -43,37 +55,98 @@ macro_rules! scope_impls {
             #[doc = "Please note that this may not work for you, as some auth flows and \"apis\" don't accept all scopes"]
             pub fn all() -> Vec<Scope> {
                 vec![
-                    $(Scope::$i,)*
+                    $(
+                        scope_impls!(@omit $(#[deprecated($depr)])* $i),
+                    )*
+                ]
+            }
+
+            #[doc = "Get a slice of all defined twitch [Scopes][Scope]."]
+            #[doc = "\n\n"]
+            #[doc = "Please note that this may not work for you, as some auth flows and \"apis\" don't accept all scopes"]
+            pub const fn all_slice() -> &'static [Scope] {
+                &[
+                    $(
+                        scope_impls!(@omit $(#[deprecated($depr)])* $i),
+                    )*
                 ]
             }
 
             #[doc = "Get a description for the token"]
-            pub fn description(&self) -> &'static str {
+            pub const fn description(&self) -> &'static str {
+                #![allow(deprecated)]
+
                 match self {
-                    $(Self::$i => $doc,)*
+
+                    $(
+                        $(#[cfg($cfg)])*
+                        Self::$i => $doc,
+                    )*
                     _ => "unknown scope"
                 }
             }
 
             #[doc = "Make a scope from a cow string"]
             pub fn parse<C>(s: C) -> Scope where C: Into<Cow<'static, str>> {
+                #![allow(deprecated)]
                 use std::borrow::Borrow;
                 let s = s.into();
                 match s.borrow() {
-                    $($rename => {Scope::$i})*,
+
+                    $(
+                        $(#[cfg($cfg)])*
+                        $rename => {Scope::$i}
+                    )*,
                     _ => Scope::Other(s)
                 }
             }
 
             /// Get the scope as a borrowed string.
             pub fn as_str(&self) -> &str {
+                #![allow(deprecated)]
                 match self {
-                    $(Scope::$i => $rename,)*
-                    Self::Other(c) =>  c.as_ref()
+                    $(
+                        $(#[cfg($cfg)])*
+                        Scope::$i => $rename,
+                    )*
+                    Self::Other(c) => c.as_ref()
+                }
+            }
+
+            /// Get the scope as a static string slice.
+            ///
+            /// ## Panics
+            ///
+            /// This function panics if the scope can't be represented as a static string slice
+            pub const fn as_static_str(&self) -> &'static str {
+                #![allow(deprecated)]
+                match self {
+                    $(
+                        $(#[cfg($cfg)])*
+                        Scope::$i => $rename,
+                    )*
+                    Self::Other(Cow::Borrowed(s)) => s,
+                    _ => panic!(),
                 }
             }
         }
-
+        #[test]
+        #[cfg(test)]
+        fn sorted() {
+            let slice = [$(
+                $(#[cfg($cfg)])*
+                $rename,
+            )*];
+            let mut slice_sorted = [$(
+                $(#[cfg($cfg)])*
+                $rename,
+            )*];
+            slice_sorted.sort();
+            for (scope, sorted) in slice.iter().zip(slice_sorted.iter()) {
+                assert_eq!(scope, sorted);
+            }
+            assert_eq!(slice, slice_sorted);
+        }
     };
 }
 
@@ -103,7 +176,8 @@ scope_impls!(
     ChannelReadStreamKey,           scope: "channel:read:stream_key",           doc: "View an authorized user’s stream key.";
     ChannelReadSubscriptions,       scope: "channel:read:subscriptions",        doc: "View a list of all subscribers to a channel and check if a user is subscribed to a channel.";
     ChannelReadVips,                scope: "channel:read:vips",                 doc: "Read the list of VIPs in your channel.";
-    ChannelSubscriptions,           scope: "channel_subscriptions",             doc: "\\[DEPRECATED\\] Read all subscribers to your channel.";
+    #[deprecated(note = "Use `ChannelReadSubscriptions` (`channel:read:subscriptions`) instead")]
+    ChannelSubscriptions,           scope: "channel_subscriptions",             doc: "Read all subscribers to your channel.";
     ChatEdit,                       scope: "chat:edit",                         doc: "Send live stream chat and rooms messages.";
     ChatRead,                       scope: "chat:read",                         doc: "View live stream chat and rooms messages.";
     ClipsEdit,                      scope: "clips:edit",                        doc: "Manage Clips for a channel.";
@@ -115,14 +189,18 @@ scope_impls!(
     ModeratorManageBlockedTerms,    scope: "moderator:manage:blocked_terms",    doc: "Manage a broadcaster’s list of blocked terms.";
     ModeratorManageChatMessages,    scope: "moderator:manage:chat_messages",    doc: "Delete chat messages in channels where you have the moderator role";
     ModeratorManageChatSettings,    scope: "moderator:manage:chat_settings",    doc: "View a broadcaster’s chat room settings.";
+    ModeratorManageShieldMode,      scope: "moderator:manage:shield_mode",      doc: "Manage a broadcaster’s Shield Mode status.";
+    ModeratorManageShoutouts,       scope: "moderator:manage:shoutouts",        doc: "Manage a broadcaster’s shoutouts.";
     ModeratorReadAutomodSettings,   scope: "moderator:read:automod_settings",   doc: "View a broadcaster’s AutoMod settings.";
     ModeratorReadBlockedTerms,      scope: "moderator:read:blocked_terms",      doc: "View a broadcaster’s list of blocked terms.";
     ModeratorReadChatSettings,      scope: "moderator:read:chat_settings",      doc: "View a broadcaster’s chat room settings.";
     ModeratorReadChatters,          scope: "moderator:read:chatters",           doc: "View the chatters in a broadcaster’s chat room.";
+    ModeratorReadFollowers,         scope: "moderator:read:followers",          doc: "Read the followers of a broadcaster.";
     ModeratorReadShieldMode,        scope: "moderator:read:shield_mode",        doc: "View a broadcaster’s Shield Mode status.";
-    ModeratorManageShieldMode,      scope: "moderator:manage:shield_mode",      doc: "Manage a broadcaster’s Shield Mode status.";
+    ModeratorReadShoutouts,         scope: "moderator:read:shoutouts",          doc: "View a broadcaster’s shoutouts.";
     UserEdit,                       scope: "user:edit",                         doc: "Manage a user object.";
     UserEditBroadcast,              scope: "user:edit:broadcast",               doc: "Edit your channel's broadcast configuration, including extension configuration. (This scope implies user:read:broadcast capability.)";
+    #[deprecated(note = "Not used anymore, see https://discuss.dev.twitch.tv/t/deprecation-of-create-and-delete-follows-api-endpoints/32351")]
     UserEditFollows,                scope: "user:edit:follows",                 doc: "\\[DEPRECATED\\] Was previously used for “Create User Follows” and “Delete User Follows.";
     UserManageBlockedUsers,         scope: "user:manage:blocked_users",         doc: "Manage the block list of a user.";
     UserManageChatColor,            scope: "user:manage:chat_color",            doc: "Update the color used for the user’s name in chat.Update User Chat Color";
@@ -163,6 +241,14 @@ mod tests {
     fn roundabout() {
         for scope in Scope::all() {
             assert_eq!(scope, Scope::parse(scope.to_string()))
+        }
+    }
+
+    #[test]
+    #[allow(deprecated)]
+    fn no_deprecated() {
+        for scope in Scope::all() {
+            assert!(scope != Scope::ChannelSubscriptions)
         }
     }
 }